requirements.txt
requests==2.31.0
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
pymysql==1.1.0
python-dotenv==1.0.1
Flask==3.0.3
